// EscalafonFrontend/src/app/interfaces/tipo-documento-identificacion.ts
export interface TipoDocumentoIdentificacion {
    /**
     * @type { number } - iTipoDocIdenId
     */
    iTipoDocIdenId?: number;

    /**
     * @type { string } - cTipoDocIdenNombre
     */
    cTipoDocIdenNombre?: string;

    /**
     * @type { number } - iTipoIdentId
     */
    iTipoIdentId?: number;

}