// EscalafonFrontend/src/app/interfaces/condicion-laboral.ts
export interface CondicionLaboral {
    /**
     * @type { number } - iCondLabId
     */
    iCondLabId?: number;

    /**
     * @type { number } - iGrupoPrestadorId
     */
    iGrupoPrestadorId?: number;

    /**
     * @type { string } - cCondLabNombre
     */
    cCondLabNombre?: string;

    /**
     * @type { string } - cCondLabCodigoSunat
     */
    cCondLabCodigoSunat?: string;

    /**
     * @type { string } - cCondLabCodigoAirshp
     */
    cCondLabCodigoAirshp?: string;

}