// EscalafonFrontend/src/app/interfaces/v-condicion-laboral-situacion-laboral.ts
export interface VCondicionLaboralSituacionLaboral {
    /**
     * @type { number } - iCondSitId
     */
    iCondSitId?: number;

    /**
     * @type { number } - iCondLabId
     */
    iCondLabId?: number;

    /**
     * @type { number } - iSitLabId
     */
    iSitLabId?: number;

    /**
     * @type { string } - cSitLabNombre
     */
    cSitLabNombre?: string;

    /**
     * @type { string } - cCondLabNombre
     */
    cCondLabNombre?: string;

    /**
     * @type { number } - Expr1
     */
    Expr1?: number;

}