// EscalafonFrontend/src/app/interfaces/v-centro-laboral.ts
export interface VCentroLaboral {
    /**
     * @type { number } - iCentLabId
     */
    iCentLabId?: number;

    /**
     * @type { number } - iDirRegId
     */
    iDirRegId?: number;

    /**
     * @type { number } - iInstGeEduId
     */
    iInstGeEduId?: number;

    /**
     * @type { number } - iModEduId
     */
    iModEduId?: number;

    /**
     * @type { boolean } - bCentLabEsVigente
     */
    bCentLabEsVigente?: boolean;

    /**
     * @type { string } - cCentLabNombre
     */
    cCentLabNombre?: string;

    /**
     * @type { string } - cCentLabCodigoModular
     */
    cCentLabCodigoModular?: string;

    /**
     * @type { number } - iNivEduId
     */
    iNivEduId?: number;

    /**
     * @type { string } - iNivEduNombre
     */
    iNivEduNombre?: string;

    /**
     * @type { string } - cDirRegNombre
     */
    cDirRegNombre?: string;

    /**
     * @type { string } - cInstGeEduNombre
     */
    cInstGeEduNombre?: string;

    /**
     * @type { string } - cModEduNombre
     */
    cModEduNombre?: string;

}